EF Splitter, MS Word Footnote conversion
The EF splitter has been rewritten to work with a JSON object:

var WET_EF_Splitter_Pairs = [
{'en': 'field-name-english','fr': 'field-name-french'},
];

for mapping fields in Drupal. This allows more reliable control with
matching fields.

The footnote conversion replaces standard footnotes with the English or
French labelling. The language is set automatically if there is a
mapping in use for the editor ids (see above) or must be selected from a
menu if no language is set for the editor.